Funcionalidad notificar dispositivos
Se notifica a los dispositivos que hayan marcado como favorito a un restaurante. Se le notifica cuando el restaurante haya creado una nueva oferta

+++ FRONTEND/lib/widgets/button_favorite.dart
@@ -3,8 +3,9 @@ import 'package:integrador/routes/imports.dart';
 // ignore: must_be_immutable
 class ButtonFavorite extends StatefulWidget {
   String? idRestaurant;
+  final Function(bool) setDoingFetch;
 
-  ButtonFavorite({super.key, this.idRestaurant});
+  ButtonFavorite({super.key, this.idRestaurant, required this.setDoingFetch});
 
   @override
   State<ButtonFavorite> createState() => _ButtonFavoriteState();
@@ -44,19 +45,54 @@ class _ButtonFavoriteState extends State<ButtonFavorite> {
     if (idUser == 'default' || idRestaurant == 'default') {
       return;
     }
+    setState(() {
+      widget.setDoingFetch(true);
+    });
     bool exists = await SQLiteDB.existsFavorite(idRestaurant);
     if (exists) {
-      await SQLiteDB.delete(idRestaurant);
-      setState(() {
-        heartColor = Colors.blueGrey;
-      });
+      await APIService.removeFavorite(idRestaurant).then((value) => {
+            if (value)
+              {
+                CustomShowDialog.make(
+                    context, "Éxito", "Se eliminó el restaurante de favoritos"),
+                setState(() {
+                  heartColor = Colors.blueGrey;
+                  widget.setDoingFetch(false);
+                }),
+                SQLiteDB.delete(idRestaurant)
+              }
+            else
+              {
+                CustomShowDialog.make(context, "Error",
+                    "No se pudo eliminar el restaurante de favoritos"),
+                setState(() {
+                  widget.setDoingFetch(false);
+                })
+              }
+          });
     } else {
       Restaurant? restaurant = await APIService.getRestaurantById(idRestaurant);
       if (restaurant != null) {
-        await SQLiteDB.insert(restaurant);
-        setState(() {
-          heartColor = Colors.red;
-        });
+        await APIService.addFavorite(idRestaurant).then((value) => {
+              if (value)
+                {
+                  CustomShowDialog.make(
+                      context, "Éxito", "Se agrego el restaurante a favoritos"),
+                  setState(() {
+                    heartColor = Colors.red;
+                    widget.setDoingFetch(false);
+                  }),
+                  SQLiteDB.insert(restaurant)
+                }
+              else
+                {
+                  CustomShowDialog.make(context, "Error",
+                      "No se pudo agregar el restaurante a favoritos"),
+                  setState(() {
+                    widget.setDoingFetch(false);
+                  })
+                }
+            });
       }
     }
   }
@@ -64,7 +100,33 @@ class _ButtonFavoriteState extends State<ButtonFavorite> {
   @override
   Widget build(BuildContext context) {
     return IconButton(
-      onPressed: () => addFavorite(widget.idRestaurant ?? 'default'),
+      onPressed: () {
+        showDialog(
+          context: context,
+          builder: (BuildContext context) {
+            return AlertDialog(
+              title: const Text('Confirmación agregar a favoritos'),
+              content: const Text(
+                  '¿Estas seguro de agregar el restaurante a favoritos?'),
+              actions: <Widget>[
+                TextButton(
+                  child: const Text('Cancelar'),
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                  },
+                ),
+                TextButton(
+                  child: const Text('Aceptar'),
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                    addFavorite(widget.idRestaurant ?? 'default');
+                  },
+                ),
+              ],
+            );
+          },
+        );
+      },
       icon: const Icon(Icons.favorite),
       color: heartColor,
       padding: const EdgeInsets.only(left: 0),
